General improvements and bug fixes
Line 4: Made the script exit with error code 1 in the case that `cd` fails.
Line 19: The variable `$UNAME` is undefined on most shells, using the command `uname` is an easy fix. Also `==` doesn't exist in posix shell (which I assume is the goal based off of the shebang, and if not it still makes it more portable).
Lines 20-24: Exporting on a separate line avoids the return value of `pwd` from being ignored.
Lines 29/31: `&>/dev/null` doesn't work in posix shell, `>/dev/null 2>&1` works though and is more portable.